requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
